MINOR: Move TestSslUtils to client test fixtures
This change moves TestSslUtils, TestUtils and accessory classes from the
test classpath in clients to test fixtures, allowing other modules to
use these classes for testing without having to include all of the test
classpath for clients.

For further details on the context and purpose of this change, check
these threads:
https://github.com/apache/kafka/pull/20376#discussion_r3167281569
https://github.com/apache/kafka/pull/22193#discussion_r3178891141
