+++ clients/src/testFixtures/java/org/apache/kafka/test/NoRetryException.java
