+++ clients/src/testFixtures/java/org/apache/kafka/test/TestCondition.java
