+++ clients/src/testFixtures/java/org/apache/kafka/test/TestUtils.java
@@ -28,12 +28,9 @@ import org.apache.kafka.common.feature.SupportedVersionRange;
 import org.apache.kafka.common.message.ApiMessageType;
 import org.apache.kafka.common.message.ApiVersionsResponseData;
 import org.apache.kafka.common.network.NetworkReceive;
-import org.apache.kafka.common.network.Send;
 import org.apache.kafka.common.protocol.ApiKeys;
 import org.apache.kafka.common.protocol.Errors;
-import org.apache.kafka.common.record.internal.UnalignedRecords;
 import org.apache.kafka.common.requests.ApiVersionsResponse;
-import org.apache.kafka.common.requests.ByteBufferChannel;
 import org.apache.kafka.common.requests.MetadataResponse.PartitionMetadata;
 import org.apache.kafka.common.requests.RequestHeader;
 import org.apache.kafka.common.serialization.StringDeserializer;
@@ -563,21 +560,6 @@ public class TestUtils {
         for (T item : iterable)
             list.add(item);
         return list;
-    }
-
-    public static ByteBuffer toBuffer(Send send) {
-        ByteBufferChannel channel = new ByteBufferChannel(send.size());
-        try {
-            assertEquals(send.size(), send.writeTo(channel));
-            channel.close();
-            return channel.buffer();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public static ByteBuffer toBuffer(UnalignedRecords records) {
-        return toBuffer(records.toSend());
     }
 
     /**
