+++ clients/src/testFixtures/java/org/apache/kafka/test/ValuelessCallable.java
